requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
pandas
openpyxl
xlrd
python-dotenv